refactor: Product 검증 추가, 패키지 변경
