+++ src/main/java/com/example/surveyapp/domain/product/service/ProductService.java
@@ -6,7 +6,7 @@ import com.example.surveyapp.domain.product.controller.dto.ProductResponseDto;
 import com.example.surveyapp.domain.product.controller.dto.ProductUpdateRequestDto;
 import com.example.surveyapp.domain.product.domain.model.Product;
 import com.example.surveyapp.domain.product.domain.model.Status;
-import com.example.surveyapp.domain.product.domain.model.repository.ProductRepository;
+import com.example.surveyapp.domain.product.domain.repository.ProductRepository;
 import com.example.surveyapp.domain.product.service.dto.ProductUpdateResponseDto;
 import com.example.surveyapp.domain.user.domain.model.User;
 import com.example.surveyapp.domain.user.domain.model.UserRoleEnum;
@@ -92,7 +92,7 @@ public class ProductService {
         Product product = productRepository.findById(id)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_PRODUCT));
 
-        if (!product.getTitle().equals(requestDto.getTitle())) { //수정하기
+        if (!product.getTitle().equals(requestDto.getTitle())) {
             boolean onlyOne = productRepository.existsByTitleAndIsDeletedFalse(requestDto.getTitle());
             if (onlyOne){
                 throw new CustomException(ErrorCode.NOT_SAME_PRODUCT_TITLE);
