+++ src/test/java/com/example/surveyapp/domain/product/domain/repository/ProductRepoTest.java
@@ -1,17 +1,12 @@
 package com.example.surveyapp.domain.product.domain.repository;
 
 import com.example.surveyapp.config.testbase.DataJpaTestBase;
-import com.example.surveyapp.config.generator.ProductFixtureGenerator;
-import com.example.surveyapp.domain.product.controller.dto.ProductCreateResponseDto;
 import com.example.surveyapp.domain.product.domain.model.Product;
 import com.example.surveyapp.domain.product.domain.model.Status;
-import com.example.surveyapp.domain.product.domain.model.repository.ProductRepository;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.util.ReflectionTestUtils;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
